Removed heap, fixed startup delay issue, added ssd1306 initialization
The startup delay issue had to do with a problem in the port. It didn't
reset the current value of the systick timer, meaning that it had to
wait until it counted all the way down before the first tick would get
executed. It turns out that this is just about a second on this
processor running that the default speed. Resetting the CVR fixed this
issue and now delays work fine if they occur within the first second of
program start.