Added IRQHandler template and changed isr getter to constexpr
The static isr(IRQn) function was changed to a constexpr. This reduced
the code size slightly, but I'm not sure the gains will last when more
IRQHandlers are added.

The template class is a quick and easy way to make a class be able to
handle multiple ISRs at the expense of more code space.