Changed all code to C in hopes of a better memory footprint
A stack size of 120 was killer. I believe C should allow me to better
control the size of stack allocations, especially in the cause of the
autolock. The only sacrifice made here was automatic initialization
since I was able to replicate the autolock with a technique used in
avr-gcc.

Code size was only reduced slightly. Without changing the stack size,
memory usage was not reduced really at all.

Also, queues have a footprint of 80 bytes.