Added overvoltage protection circuit to +3.3V
Part selection is complete. There are things without part numbers.

It uses a TL431 as a precision comparator with integrated voltage
reference to activate a latch which both shuts down the regulator and
throws a crobar across +3.3V. There is also space for a fuse on +3.3V
(I'm thinking a PTC resettable fuse could go there, but I'm not sure
what the tradeoffs are for that).