Fixed thermal cutoff circuit and finished footprint selection
The main issue with the thermal cutoff was the open-collector maximum
voltage. I had to find a (slightly more expensive) comparator with an
output that was tolerant. Other comparators that would be instead
tolerant of being run off 12V were about the same price.

The fuse I'm not sure about. I'm just going to short it out. The
regulator can handle a short circuit and so can the crowbar. I also
couldn't decide where to put the feedback network. If I put it after the
fuse it would compensate for any voltage drop across it. However, it
would also cause the regulator to try to boost the voltage insanely high
during a short circuit across the fuse and maybe cause it to pop (I
could really only find ones rated to 6V). I instead placed it before the
fuse. I'm still not sure.