Changed method for building oval pins for gerber export
When the gerbers need to be exported, the cutouts should be moved down
from the user drawing layer. The gerbers should be exported and the
board revered to the previous state. The gerbers will now be stored in
the git repository since I've learned it's dumb not to keep a copy of
those.