Changed tasks to start leaning towards an unallocated stack
The way this is going to work is once the scheduler is initiated, the
current stack pointer will be noted. All of the configured tasks will
then have their stacks allocated in the remaining stack space. I've
decided on this method since syntactically it looks simpler.

To prevent stack overflow issues, I'd like some way to allocate fake
space in the linker file for all the stacks to ensure that everything
fits. Alternatively, I could switch over to the statically allocated
stack method that FreeRTOS uses.