Added support for adding tasks while the RTOS is running
This entailed a lot of things that kind of cascaded:

- First, I needed to add a dedicated idle task outside the array of
tasks. This way I could ensure the idle task was always the lowest
priority.
- That, in turn, meant I needed a way to say that the idle task's stack
pointer should be stored. The obvious way was to use current_task=None
to mean that the idle task is running, rather than the RTOS hasn't
started.
- Because I no longer had a way to determine if the RTOS had started, I
needed to find a different way. I decided to implement the msp/psp
switchover, with the PENDSV interrupt assuming that if it interrupted
while the processor was on the msp, then the RTOS hadn't started yet and
so no task pointer needed to be stored. If PENDSV interrupts running the
psp, then the psp is stored in the current task (which if it is none, is
the idle task).