Changed Block trait to no longer use static lifetimes
This really causes usability problems because any function receiving the
reference is free to operate directly on the Block forever. We don't
actually want that. It's like we want to force a static lifetime of the
reference, but not promise to use it. I don't really know of a good way
to enforce this, so instead I just made a UniqueId struct based on
UnsafeCell which creates an atomically incrementing UniqueId.