Fixed SysTick to be declared
I had to add the `rt` crate to board-support so that I could get the
exception! macro. I guess the interrupt! and exception! macros actually
define the function, so it gets to be `pub use`d back up at main.

The program now runs without hard faults. So difficult to figure out
what was wrong...